Updated AIMD for floating point window

+++ datagrump/controller.hh
@@ -9,7 +9,7 @@ class Controller
 {
 private:
   bool debug_; /* Enables debugging output */
-  unsigned int the_window_size;
+  double the_window_size;
   uint64_t seq_number_sent;
   uint64_t seq_number_acked;
   uint64_t num_acks;
